PackageInfoExtensions.
