+++ app/src/main/kotlin/com/absinthe/libchecker/ui/album/ComparisonActivity.kt
@@ -40,9 +40,11 @@ import com.absinthe.libchecker.ui.detail.SnapshotDetailActivity
 import com.absinthe.libchecker.ui.fragment.snapshot.TimeNodeBottomSheetDialogFragment
 import com.absinthe.libchecker.utils.FileUtils
 import com.absinthe.libchecker.utils.PackageUtils
-import com.absinthe.libchecker.utils.PackageUtils.getPermissionsList
 import com.absinthe.libchecker.utils.extensions.addPaddingTop
 import com.absinthe.libchecker.utils.extensions.dp
+import com.absinthe.libchecker.utils.extensions.getPackageSize
+import com.absinthe.libchecker.utils.extensions.getPermissionsList
+import com.absinthe.libchecker.utils.extensions.getVersionCode
 import com.absinthe.libchecker.utils.showToast
 import com.absinthe.libchecker.utils.toJson
 import com.absinthe.libchecker.view.snapshot.ComparisonDashboardView
@@ -273,9 +275,11 @@ class ComparisonActivity : BaseActivity<ActivityComparisonBinding>() {
 
           is SnapshotViewModel.Effect.DashboardCountChange -> {
             if (it.isLeft) {
-              dashboardView.container.leftPart.tvSnapshotAppsCountText.text = it.snapshotCount.toString()
+              dashboardView.container.leftPart.tvSnapshotAppsCountText.text =
+                it.snapshotCount.toString()
             } else {
-              dashboardView.container.rightPart.tvSnapshotAppsCountText.text = it.snapshotCount.toString()
+              dashboardView.container.rightPart.tvSnapshotAppsCountText.text =
+                it.snapshotCount.toString()
             }
           }
         }
@@ -427,35 +431,34 @@ class ComparisonActivity : BaseActivity<ActivityComparisonBinding>() {
       }
     }
 
-    if (pi == null) {
-      throw IllegalStateException("PackageInfo is null")
-    }
-    val ai = pi!!.applicationInfo
-    return SnapshotItem(
-      id = null,
-      packageName = pi!!.packageName,
-      timeStamp = -1L,
-      label = ai.loadLabel(packageManager).toString(),
-      versionName = pi!!.versionName ?: "null",
-      versionCode = PackageUtils.getVersionCode(pi!!),
-      installedTime = pi!!.firstInstallTime,
-      lastUpdatedTime = pi!!.lastUpdateTime,
-      isSystem = (ai.flags and ApplicationInfo.FLAG_SYSTEM) == ApplicationInfo.FLAG_SYSTEM,
-      abi = PackageUtils.getAbi(pi!!).toShort(),
-      targetApi = ai.targetSdkVersion.toShort(),
-      nativeLibs = PackageUtils.getNativeDirLibs(pi!!).toJson().orEmpty(),
-      services = PackageUtils.getComponentStringList(pi!!.packageName, SERVICE, false)
-        .toJson().orEmpty(),
-      activities = PackageUtils.getComponentStringList(pi!!.packageName, ACTIVITY, false)
-        .toJson().orEmpty(),
-      receivers = PackageUtils.getComponentStringList(pi!!.packageName, RECEIVER, false)
-        .toJson().orEmpty(),
-      providers = PackageUtils.getComponentStringList(pi!!.packageName, PROVIDER, false)
-        .toJson().orEmpty(),
-      permissions = pi!!.getPermissionsList().toJson().orEmpty(),
-      metadata = PackageUtils.getMetaDataItems(pi!!).toJson().orEmpty(),
-      packageSize = PackageUtils.getPackageSize(pi!!, true)
-    )
+    pi?.let {
+      val ai = it.applicationInfo
+      return SnapshotItem(
+        id = null,
+        packageName = it.packageName,
+        timeStamp = -1L,
+        label = ai.loadLabel(packageManager).toString(),
+        versionName = it.versionName ?: "null",
+        versionCode = it.getVersionCode(),
+        installedTime = it.firstInstallTime,
+        lastUpdatedTime = it.lastUpdateTime,
+        isSystem = (ai.flags and ApplicationInfo.FLAG_SYSTEM) == ApplicationInfo.FLAG_SYSTEM,
+        abi = PackageUtils.getAbi(it).toShort(),
+        targetApi = ai.targetSdkVersion.toShort(),
+        nativeLibs = PackageUtils.getNativeDirLibs(it).toJson().orEmpty(),
+        services = PackageUtils.getComponentStringList(it.packageName, SERVICE, false)
+          .toJson().orEmpty(),
+        activities = PackageUtils.getComponentStringList(it.packageName, ACTIVITY, false)
+          .toJson().orEmpty(),
+        receivers = PackageUtils.getComponentStringList(it.packageName, RECEIVER, false)
+          .toJson().orEmpty(),
+        providers = PackageUtils.getComponentStringList(it.packageName, PROVIDER, false)
+          .toJson().orEmpty(),
+        permissions = it.getPermissionsList().toJson().orEmpty(),
+        metadata = PackageUtils.getMetaDataItems(it).toJson().orEmpty(),
+        packageSize = it.getPackageSize(true)
+      )
+    } ?: throw IllegalStateException("PackageInfo is null")
   }
 
   private fun getSuitableLayoutManager(): RecyclerView.LayoutManager {
@@ -465,6 +468,7 @@ class ComparisonActivity : BaseActivity<ActivityComparisonBinding>() {
         2,
         StaggeredGridLayoutManager.VERTICAL
       )
+
       else -> throw IllegalStateException("Wrong orientation at AppListFragment.")
     }
   }
