+++ app/src/main/kotlin/com/absinthe/libchecker/utils/PackageUtils.kt
@@ -6,16 +6,13 @@ import android.content.pm.ApplicationInfo
 import android.content.pm.ApplicationInfoHidden
 import android.content.pm.ComponentInfo
 import android.content.pm.PackageInfo
-import android.content.pm.PackageInfoHidden
 import android.content.pm.PackageManager
 import android.content.pm.Signature
-import android.content.res.Configuration
 import android.os.Build
 import android.os.Process
 import android.text.format.Formatter
 import androidx.annotation.DrawableRes
 import androidx.collection.arrayMapOf
-import androidx.core.content.pm.PackageInfoCompat
 import androidx.core.text.isDigitsOnly
 import com.absinthe.libchecker.LibCheckerApp
 import com.absinthe.libchecker.R
@@ -50,20 +47,21 @@ import com.absinthe.libchecker.constant.Constants.X86
 import com.absinthe.libchecker.constant.Constants.X86_64
 import com.absinthe.libchecker.constant.Constants.X86_64_STRING
 import com.absinthe.libchecker.constant.Constants.X86_STRING
-import com.absinthe.libchecker.database.entity.Features
-import com.absinthe.libchecker.model.KotlinToolingMetadata
 import com.absinthe.libchecker.model.LibStringItem
 import com.absinthe.libchecker.model.StatefulComponent
 import com.absinthe.libchecker.utils.dex.DexLibMap
 import com.absinthe.libchecker.utils.dex.FastDexFileFactory
 import com.absinthe.libchecker.utils.elf.ELFParser
+import com.absinthe.libchecker.utils.extensions.getPermissionsList
+import com.absinthe.libchecker.utils.extensions.getStatefulPermissionsList
+import com.absinthe.libchecker.utils.extensions.getVersionCode
+import com.absinthe.libchecker.utils.extensions.isOverlay
+import com.absinthe.libchecker.utils.extensions.isUse32BitAbi
 import com.absinthe.libchecker.utils.extensions.md5
 import com.absinthe.libchecker.utils.extensions.sha1
 import com.absinthe.libchecker.utils.extensions.sha256
 import com.absinthe.libchecker.utils.extensions.toClassDefType
 import com.absinthe.libchecker.utils.extensions.toHexString
-import com.absinthe.libchecker.utils.manifest.HiddenPermissionsReader
-import com.absinthe.libchecker.utils.manifest.ManifestReader
 import com.absinthe.libchecker.utils.manifest.StaticLibraryReader
 import dev.rikka.tools.refine.Refine
 import java.io.File
@@ -71,16 +69,10 @@ import java.io.InputStream
 import java.security.interfaces.DSAPublicKey
 import java.security.interfaces.RSAPublicKey
 import java.text.DateFormat
-import java.util.Properties
 import java.util.zip.ZipEntry
 import java.util.zip.ZipFile
 import javax.security.cert.X509Certificate
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import okio.buffer
-import okio.source
 import org.jf.dexlib2.Opcodes
-import rikka.material.app.LocaleDelegate
 import timber.log.Timber
 
 object PackageUtils {
@@ -124,33 +116,11 @@ object PackageUtils {
 
   /**
    * Get version code of an app
-   * @param packageInfo PackageInfo
-   * @return version code as Long Integer
-   */
-  fun getVersionCode(packageInfo: PackageInfo): Long {
-    return PackageInfoCompat.getLongVersionCode(packageInfo)
-  }
-
-  /**
-   * Get version code of an app
    * @param packageName packageName
    * @return version code as Long Integer
    */
   fun getVersionCode(packageName: String): Long {
-    return PackageInfoCompat.getLongVersionCode(getPackageInfo(packageName))
-  }
-
-  /**
-   * Get version string of an app ( 1.0.0(1) )
-   * @param packageInfo PackageInfo
-   * @return version code as String
-   */
-  fun getVersionString(packageInfo: PackageInfo): String {
-    return try {
-      "${packageInfo.versionName ?: "<unknown>"} (${getVersionCode(packageInfo)})"
-    } catch (e: PackageManager.NameNotFoundException) {
-      "Unknown"
-    }
+    return getPackageInfo(packageName).getVersionCode()
   }
 
   /**
@@ -161,51 +131,6 @@ object PackageUtils {
    */
   fun getVersionString(versionName: String, versionCode: Long): String {
     return "$versionName ($versionCode)"
-  }
-
-  /**
-   * Get target api string of an app ( API 30 )
-   * @param packageInfo PackageInfo
-   * @return version code as String
-   */
-  fun getTargetApiString(packageInfo: PackageInfo): String {
-    return runCatching {
-      packageInfo.applicationInfo.targetSdkVersion.toString()
-    }.getOrDefault("?")
-  }
-
-  private const val compileSdkVersion = "compileSdkVersion"
-
-  /**
-   * Get minSdkVersion of an app
-   * @param packageInfo PackageInfo
-   * @return minSdkVersion
-   */
-  fun getMinSdkVersion(packageInfo: PackageInfo): Int = packageInfo.applicationInfo.minSdkVersion
-
-  /**
-   * Get compileSdkVersion of an app
-   * @param packageInfo PackageInfo
-   * @return compileSdkVersion
-   */
-  fun getCompileSdkVersion(packageInfo: PackageInfo): String {
-    return runCatching {
-      val version = if (OsUtils.atLeastS()) {
-        packageInfo.applicationInfo.compileSdkVersion
-      } else {
-        val demands = ManifestReader.getManifestProperties(
-          File(packageInfo.applicationInfo.sourceDir),
-          arrayOf(compileSdkVersion)
-        )
-        demands[compileSdkVersion]?.toString()?.toInt() ?: 0
-      }
-
-      if (version == 0) {
-        "?"
-      } else {
-        version.toString()
-      }
-    }.getOrDefault("?")
   }
 
   /**
@@ -381,32 +306,6 @@ object PackageUtils {
     return packageInfo.applicationInfo.splitSourceDirs
   }
 
-  /**
-   * Check if an app uses split apks
-   * @return true if it uses split apks
-   */
-  fun PackageInfo.isSplitsApk(): Boolean {
-    return !applicationInfo.splitSourceDirs.isNullOrEmpty()
-  }
-
-  /**
-   * Check if an app uses Kotlin language
-   * @return true if it uses Kotlin language
-   */
-  fun PackageInfo.isKotlinUsed(): Boolean {
-    return runCatching {
-      val file = File(applicationInfo.sourceDir)
-
-      ZipFile(file).use {
-        it.getEntry("kotlin-tooling-metadata.json") != null ||
-          it.getEntry("kotlin/kotlin.kotlin_builtins") != null ||
-          it.getEntry("META-INF/services/kotlinx.coroutines.CoroutineExceptionHandler") != null ||
-          it.getEntry("META-INF/services/kotlinx.coroutines.internal.MainDispatcherFactory") != null ||
-          isKotlinUsedInClassDex(file)
-      }
-    }.getOrDefault(false)
-  }
-
   const val STATIC_LIBRARY_SOURCE_PREFIX = "[Path] "
   const val VERSION_CODE_PREFIX = "[Version Code] "
 
@@ -482,7 +381,7 @@ object PackageUtils {
    * @param file APK file of the app
    * @return true if it uses Kotlin language
    */
-  private fun isKotlinUsedInClassDex(file: File): Boolean {
+  fun isKotlinUsedInClassDex(file: File): Boolean {
     return findDexClasses(
       file,
       listOf("kotlin.*".toClassDefType(), "kotlinx.*".toClassDefType())
@@ -1134,10 +1033,6 @@ object PackageUtils {
     }.getOrElse { emptyList() }
   }
 
-  fun PackageInfo.getPermissionsList(): List<String> {
-    return requestedPermissions?.toList() ?: emptyList()
-  }
-
   /**
    * Get permissions of an application with granted state
    * @param packageName Package name of the app
@@ -1149,33 +1044,6 @@ object PackageUtils {
     }.getOrElse { emptyList() }
   }
 
-  fun PackageInfo.getStatefulPermissionsList(): List<Pair<String, Boolean>> {
-    val flags = requestedPermissionsFlags
-    val hidden = HiddenPermissionsReader.getHiddenPermissions(File(applicationInfo.sourceDir)).filter { (_, v) ->
-      OsUtils.higherThan(v as Int)
-    }
-
-    if (flags?.size != requestedPermissions?.size) {
-      return requestedPermissions?.map { it to true }?.toMutableList()?.apply {
-        if (hidden.isNotEmpty()) {
-          hidden.forEach { (p, v) ->
-            add("$p (maxSdkVersion: $v)" to false)
-          }
-        }
-      } ?: emptyList()
-    }
-
-    return requestedPermissions?.mapIndexed { index, s ->
-      s to (flags[index] and PackageInfo.REQUESTED_PERMISSION_GRANTED != 0)
-    }?.toMutableList()?.apply {
-      if (hidden.isNotEmpty()) {
-        hidden.forEach { (p, v) ->
-          add("$p (maxSdkVersion: $v)" to false)
-        }
-      }
-    } ?: emptyList()
-  }
-
   /**
    * Check if an app is installed
    * @return true if it is installed
@@ -1184,208 +1052,6 @@ object PackageUtils {
     return runCatching {
       PackageManagerCompat.getApplicationInfo(pkgName, 0).enabled
     }.getOrDefault(false)
-  }
-
-  private const val AGP_KEYWORD = "androidGradlePluginVersion"
-  private const val AGP_KEYWORD2 = "Created-By: Android Gradle "
-
-  /**
-   * Get Android Gradle Plugin version of an app
-   * @param packageInfo PackageInfo
-   * @return Android Gradle Plugin version
-   */
-  fun getAGPVersion(packageInfo: PackageInfo): String? {
-    runCatching {
-      ZipFile(File(packageInfo.applicationInfo.sourceDir)).use { zipFile ->
-        zipFile.getEntry("META-INF/com/android/build/gradle/app-metadata.properties")?.let { ze ->
-          Properties().apply {
-            load(zipFile.getInputStream(ze))
-            getProperty(AGP_KEYWORD)?.let {
-              return it
-            }
-          }
-        }
-        zipFile.getEntry("META-INF/MANIFEST.MF")?.let { ze ->
-          zipFile.getInputStream(ze).source().buffer().use {
-            while (true) {
-              it.readUtf8Line()?.let { line ->
-                if (line.startsWith(AGP_KEYWORD2)) {
-                  return line.removePrefix(AGP_KEYWORD2)
-                }
-              } ?: break
-            }
-          }
-        }
-        arrayOf(
-          "META-INF/androidx.databinding_viewbinding.version",
-          "META-INF/androidx.databinding_databindingKtx.version",
-          "META-INF/androidx.databinding_library.version"
-        ).forEach { entry ->
-          zipFile.getEntry(entry)?.let { ze ->
-            zipFile.getInputStream(ze).source().buffer().use { bs ->
-              return bs.readUtf8Line().takeIf { it?.isNotBlank() == true }
-            }
-          }
-        }
-      }
-    }
-
-    return null
-  }
-
-  fun getPackageSize(packageInfo: PackageInfo, includeSplits: Boolean): Long {
-    var size: Long = FileUtils.getFileSize(packageInfo.applicationInfo.sourceDir)
-
-    if (!includeSplits) {
-      return size
-    }
-
-    getSplitsSourceDir(packageInfo)?.forEach {
-      runCatching {
-        size += FileUtils.getFileSize(it)
-      }
-    }
-    return size
-  }
-
-  fun PackageInfo.getFeatures(): Int {
-    var features = 0
-    val resultList = findDexClasses(
-      File(applicationInfo.sourceDir),
-      listOf(
-        "androidx.compose.*".toClassDefType(),
-        "rx.*".toClassDefType(),
-        "io.reactivex.*".toClassDefType(),
-        "io.reactivex.rxjava3.*".toClassDefType(),
-        "io.reactivex.rxjava3.kotlin.*".toClassDefType(),
-        "io.reactivex.rxkotlin".toClassDefType(),
-        "rx.lang.kotlin".toClassDefType(),
-        "io.reactivex.rxjava3.android.*".toClassDefType(),
-        "io.reactivex.android.*".toClassDefType(),
-        "rx.android.*".toClassDefType()
-      )
-    )
-    if (isSplitsApk()) {
-      features = features or Features.SPLIT_APKS
-    }
-    if (isKotlinUsed()) {
-      features = features or Features.KOTLIN_USED
-    }
-    if (getAGPVersion(this)?.isNotBlank() == true) {
-      features = features or Features.AGP
-    }
-    if (isXposedModule()) {
-      features = features or Features.XPOSED_MODULE
-    }
-    if (isPlayAppSigning()) {
-      features = features or Features.PLAY_SIGNING
-    }
-    if (isPWA()) {
-      features = features or Features.PWA
-    }
-    if (isUseJetpackCompose(resultList)) {
-      features = features or Features.JETPACK_COMPOSE
-    }
-    if (isRxJavaUsed(resultList)) {
-      features = features or Features.RX_JAVA
-    }
-    if (isRxKotlinUsed(resultList)) {
-      features = features or Features.RX_KOTLIN
-    }
-    if (isRxAndroidUsed(resultList)) {
-      features = features or Features.RX_ANDROID
-    }
-
-    return features
-  }
-
-  fun PackageInfo.isXposedModule(): Boolean {
-    return applicationInfo.metaData?.getBoolean("xposedmodule") == true ||
-      applicationInfo.metaData?.containsKey("xposedminversion") == true
-  }
-
-  fun PackageInfo.isPlayAppSigning(): Boolean {
-    return applicationInfo.metaData?.getString("com.android.stamp.type") == "STAMP_TYPE_DISTRIBUTION_APK" &&
-      applicationInfo.metaData?.getString("com.android.stamp.source") == "https://play.google.com/store"
-  }
-
-  fun PackageInfo.isPWA(): Boolean {
-    return applicationInfo.metaData?.keySet()
-      ?.any { it.startsWith("org.chromium.webapk.shell_apk") } == true
-  }
-
-  fun PackageInfo.isOverlay(): Boolean {
-    return try {
-      Refine.unsafeCast<PackageInfoHidden>(this).isOverlayPackage
-    } catch (t: Throwable) {
-      if (applicationInfo.sourceDir == null) return false
-      val demands =
-        ManifestReader.getManifestProperties(File(applicationInfo.sourceDir), arrayOf("overlay"))
-      return demands["overlay"] as? Boolean ?: false
-    }
-  }
-
-  fun ApplicationInfo.isUse32BitAbi(): Boolean {
-    runCatching {
-      val demands = ManifestReader.getManifestProperties(File(sourceDir), arrayOf("use32bitAbi"))
-      return demands["use32bitAbi"] as? Boolean ?: false
-    }.getOrNull() ?: return false
-  }
-
-  fun getKotlinPluginVersion(packageInfo: PackageInfo): String? {
-    return runCatching {
-      ZipFile(packageInfo.applicationInfo.sourceDir).use { zip ->
-        val entry = zip.getEntry("kotlin-tooling-metadata.json") ?: return@runCatching null
-        zip.getInputStream(entry).source().buffer().use {
-          val json = it.readUtf8().fromJson<KotlinToolingMetadata>()
-          return json?.buildPluginVersion.takeIf { json?.buildPlugin == "org.jetbrains.kotlin.gradle.plugin.KotlinAndroidPluginWrapper" }
-        }
-      }
-    }.getOrNull()
-  }
-
-  fun PackageInfo.isUseJetpackCompose(foundList: List<String>? = null): Boolean {
-    val usedInMetaInf = runCatching {
-      val file = File(applicationInfo.sourceDir)
-
-      ZipFile(file).use {
-        it.entries().asSequence().any { entry ->
-          entry.isDirectory.not() &&
-            entry.name.startsWith("androidx.compose") &&
-            entry.name.endsWith(".version")
-        }
-      }
-    }.getOrDefault(false)
-    if (usedInMetaInf) {
-      return true
-    }
-    if (foundList.isNullOrEmpty().not()) {
-      return foundList?.contains("androidx.compose.*".toClassDefType()) == true
-    }
-    return findDexClasses(
-      File(applicationInfo.sourceDir),
-      listOf("androidx.compose.*".toClassDefType())
-    ).isNotEmpty()
-  }
-
-  fun getJetpackComposeVersion(packageInfo: PackageInfo): String? {
-    runCatching {
-      ZipFile(File(packageInfo.applicationInfo.sourceDir)).use { zipFile ->
-        arrayOf(
-          "META-INF/androidx.compose.runtime_runtime.version",
-          "META-INF/androidx.compose.ui_ui.version",
-          "META-INF/androidx.compose.ui_ui-tooling-preview.version"
-        ).forEach { entry ->
-          zipFile.getEntry(entry)?.let { ze ->
-            zipFile.getInputStream(ze).source().buffer().use { bs ->
-              return bs.readUtf8Line().takeIf { it?.isNotBlank() == true }
-            }
-          }
-        }
-      }
-    }
-
-    return null
   }
 
   private fun getElfType(file: File): Int {
@@ -1400,222 +1066,7 @@ object PackageUtils {
     return ELFParser(input)
   }
 
-  private const val RX_MAJOR_ONE = "1"
-  private const val RX_MAJOR_TWO = "2"
-  private const val RX_MAJOR_THREE = "3"
-
-  /**
-   * Check if an app uses RxJava framework
-   * @return true if it uses RxJava framework
-   */
-  fun PackageInfo.isRxJavaUsed(foundList: List<String>? = null): Boolean {
-    val usedInMetaInf = runCatching {
-      val file = File(applicationInfo.sourceDir)
-
-      ZipFile(file).use {
-        it.getEntry("META-INF/rxjava.properties") != null
-      }
-    }.getOrDefault(false)
-    if (usedInMetaInf) {
-      return true
-    }
-    if (foundList.isNullOrEmpty().not()) {
-      return foundList?.contains("rx.*".toClassDefType()) == true ||
-        foundList?.contains("io.reactivex.*".toClassDefType()) == true ||
-        foundList?.contains("io.reactivex.rxjava3.*".toClassDefType()) == true
-    }
-    return findDexClasses(
-      File(applicationInfo.sourceDir),
-      listOf(
-        "rx.*".toClassDefType(),
-        "io.reactivex.*".toClassDefType(),
-        "io.reactivex.rxjava3.*".toClassDefType()
-      ),
-      hasAny = true
-    ).isNotEmpty()
-  }
-
-  private const val REACTIVEX_KEYWORD = "Implementation-Version"
-
-  suspend fun getRxJavaVersion(packageInfo: PackageInfo): String? {
-    return withContext(Dispatchers.IO) {
-      runCatching {
-        ZipFile(File(packageInfo.applicationInfo.sourceDir)).use { zipFile ->
-          zipFile.getEntry("META-INF/rxjava.properties")?.let { ze ->
-            Properties().apply {
-              load(zipFile.getInputStream(ze))
-              getProperty(REACTIVEX_KEYWORD)?.let {
-                return@withContext it
-              }
-            }
-          }
-        }
-        val resultList = findDexClasses(
-          File(packageInfo.applicationInfo.sourceDir),
-          listOf(
-            "rx.*".toClassDefType(),
-            "io.reactivex.*".toClassDefType(),
-            "io.reactivex.rxjava3.*".toClassDefType()
-          ),
-          hasAny = true
-        )
-        if (resultList.contains("io.reactivex.rxjava3.*".toClassDefType())) {
-          return@withContext RX_MAJOR_THREE
-        }
-        if (resultList.contains("io.reactivex.*".toClassDefType())) {
-          return@withContext RX_MAJOR_TWO
-        }
-        if (resultList.contains("rx.*".toClassDefType())) {
-          return@withContext RX_MAJOR_ONE
-        }
-      }
-      return@withContext null
-    }
-  }
-
-  /**
-   * Check if an app uses RxKotlin framework
-   * @return true if it uses RxKotlin framework
-   */
-  fun PackageInfo.isRxKotlinUsed(foundList: List<String>? = null): Boolean {
-    val usedInMetaInf = runCatching {
-      val file = File(applicationInfo.sourceDir)
-
-      ZipFile(file).use {
-        it.getEntry("META-INF/rxkotlin.properties") != null
-      }
-    }.getOrDefault(false)
-    if (usedInMetaInf) {
-      return true
-    }
-    if (foundList.isNullOrEmpty().not()) {
-      return foundList?.contains("io.reactivex.rxjava3.kotlin.*".toClassDefType()) == true ||
-        foundList?.contains("io.reactivex.rxkotlin".toClassDefType()) == true ||
-        foundList?.contains("rx.lang.kotlin".toClassDefType()) == true
-    }
-    return findDexClasses(
-      File(applicationInfo.sourceDir),
-      listOf(
-        "io.reactivex.rxjava3.kotlin.*".toClassDefType(),
-        "io.reactivex.rxkotlin".toClassDefType(),
-        "rx.lang.kotlin".toClassDefType()
-      ),
-      hasAny = true
-    ).isNotEmpty()
-  }
-
-  suspend fun getRxKotlinVersion(packageInfo: PackageInfo): String? {
-    return withContext(Dispatchers.IO) {
-      runCatching {
-        ZipFile(File(packageInfo.applicationInfo.sourceDir)).use { zipFile ->
-          zipFile.getEntry("META-INF/rxkotlin.properties")?.let { ze ->
-            Properties().apply {
-              load(zipFile.getInputStream(ze))
-              getProperty(REACTIVEX_KEYWORD)?.let {
-                return@withContext it
-              }
-            }
-          }
-        }
-        val resultList = findDexClasses(
-          File(packageInfo.applicationInfo.sourceDir),
-          listOf(
-            "io.reactivex.rxjava3.kotlin.*".toClassDefType(),
-            "io.reactivex.rxkotlin".toClassDefType(),
-            "rx.lang.kotlin".toClassDefType()
-          ),
-          hasAny = true
-        )
-        if (resultList.contains("io.reactivex.rxjava3.kotlin.*".toClassDefType())) {
-          return@withContext RX_MAJOR_THREE
-        }
-        if (resultList.contains("io.reactivex.rxkotlin".toClassDefType())) {
-          return@withContext RX_MAJOR_TWO
-        }
-        if (resultList.contains("rx.lang.kotlin".toClassDefType())) {
-          return@withContext RX_MAJOR_ONE
-        }
-      }
-      return@withContext null
-    }
-  }
-
-  /**
-   * Check if an app uses RxAndroid framework
-   * @return true if it uses RxAndroid framework
-   */
-  fun PackageInfo.isRxAndroidUsed(foundList: List<String>? = null): Boolean {
-    if (foundList.isNullOrEmpty().not()) {
-      return foundList?.contains("io.reactivex.rxjava3.android.*".toClassDefType()) == true ||
-        foundList?.contains("io.reactivex.android.*".toClassDefType()) == true ||
-        foundList?.contains("rx.android.*".toClassDefType()) == true
-    }
-    return findDexClasses(
-      File(applicationInfo.sourceDir),
-      listOf(
-        "io.reactivex.rxjava3.android.*".toClassDefType(),
-        "io.reactivex.android.*".toClassDefType(),
-        "rx.android.*".toClassDefType()
-      ),
-      hasAny = true
-    ).isNotEmpty()
-  }
-
-  suspend fun getRxAndroidVersion(packageInfo: PackageInfo): String? {
-    return withContext(Dispatchers.IO) {
-      val resultList = findDexClasses(
-        File(packageInfo.applicationInfo.sourceDir),
-        listOf(
-          "io.reactivex.rxjava3.android.*".toClassDefType(),
-          "io.reactivex.android.*".toClassDefType(),
-          "rx.android.*".toClassDefType()
-        ),
-        hasAny = true
-      )
-      if (resultList.contains("io.reactivex.rxjava3.android.*".toClassDefType())) {
-        return@withContext RX_MAJOR_THREE
-      }
-      if (resultList.contains("io.reactivex.android.*".toClassDefType())) {
-        return@withContext RX_MAJOR_TWO
-      }
-      if (resultList.contains("rx.android.*".toClassDefType())) {
-        return@withContext RX_MAJOR_ONE
-      }
-      return@withContext null
-    }
-  }
-
-  /**
-   * Get signatures of an app
-   * @param packageInfo PackageInfo
-   * @return List of LibStringItem
-   */
-  fun getSignatures(context: Context, packageInfo: PackageInfo): Sequence<LibStringItem> {
-    val localedContext = context.createConfigurationContext(
-      Configuration(context.resources.configuration).apply {
-        setLocale(LocaleDelegate.defaultLocale)
-      }
-    )
-    val dateFormat = DateFormat.getDateTimeInstance(
-      DateFormat.LONG,
-      DateFormat.LONG,
-      LocaleDelegate.defaultLocale
-    )
-    return if (OsUtils.atLeastP() && packageInfo.signingInfo != null) {
-      if (packageInfo.signingInfo.hasMultipleSigners()) {
-        packageInfo.signingInfo.apkContentsSigners
-      } else {
-        packageInfo.signingInfo.signingCertificateHistory
-      }
-    } else {
-      @Suppress("DEPRECATION")
-      packageInfo.signatures
-    }.orEmpty().asSequence().map {
-      describeSignature(localedContext, dateFormat, it)
-    }
-  }
-
-  private fun describeSignature(
+  fun describeSignature(
     context: Context,
     dateFormat: DateFormat,
     signature: Signature
